refactor: moved entities to domain package
