+++ src/main/java/org/example/animal/domain/Animal.java
@@ -1,4 +1,4 @@
-package org.example.animal;
+package org.example.animal.domain;
 
 public abstract class Animal {
     private AnimalType type;
