+++ src/main/java/org/example/animal/domain/AnimalType.java
@@ -1,4 +1,4 @@
-package org.example.animal;
+package org.example.animal.domain;
 
 public enum AnimalType {
     TERRESTRE,
