+++ src/main/java/org/example/animal/domain/AquaticAnimal.java
@@ -1,4 +1,4 @@
-package org.example.animal;
+package org.example.animal.domain;
 
 public class AquaticAnimal extends Animal{
     public AquaticAnimal(String onomatopoeia, String species) {
